requirements.txt
pandas
numpy
yfinance
matplotlib
plotly
scikit-learn
scipy
statsmodels
PyYAML
openpyxl
arch
streamlit
seaborn
